+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 
 ## Unreleased
+### Added
+- Add the ability to read headers from response and `FacebookRequestError`.
 
 ## v5.0.0
 ### Changed